src/pyproject.toml
[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "django-stubs>=5.0.4",
    "djangorestframework-stubs>=3.15.0",
]

[tool.pdm.scripts]
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
start = "python manage.py runserver"
dbsync.composite = [
  "makemigrations",
  "migrate",
]
post_init = { composite = ["pdm install", "migrate", "start"] }

[tool.basedpyright]
typeCheckingMode = "standard"
reportIncompatibleVariableOverride = "none"

[project]
name = "erp_backend"
version = "0.1.0"
description = "Django project template managed by PDM"
authors = [
    {name = "JuanBaut", email = "deserd@protonmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.9",
    "django~=4.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}
